+++ packages/@runlightyear/cli/src/index.ts
@@ -7,14 +7,21 @@ import { login } from "./commands/login";
 import { dev } from "./commands/dev";
 import { deploy } from "./commands/deploy";
 import { run } from "./commands/run";
+import { test } from "./commands/test";
+import { prepareConsole } from "./logging";
 
 dotenv.config();
+
+prepareConsole();
+
+console.log("XXX hi there");
 
 program
   .name("lightyear")
   .description("Lightyear CLI")
   .version(packageJson.version)
   .addCommand(create)
+  .addCommand(test)
   .addCommand(login)
   .addCommand(dev)
   .addCommand(deploy)